Add files via upload
updates doc-strings and typing

+++ scale_objects.py
@@ -12,11 +12,12 @@ CHARACTER_WIDTH: int = round(s.WIDTH * 0.047)
 CHARACTER_HEIGHT: int = round(s.HEIGHT * 0.094)
 CHARACTER_POS_X: float = s.WIDTH / 2
 CHARACTER_POS_Y: float = s.HEIGHT - (s.HEIGHT * 0.25)
-CHARACTER_MASS: int = 1.2
+CHARACTER_MASS: float = 1.2
 
 CHARACTER_MAX_VELOCITY_X: int = 2000
 
 CHARACTER_JUMP_FORCE: int = 1200
+# Formula for calculating the jump height based on the characters speed
 CHARACTER_JUMP_VELOCITY_FACTOR: float = (4000 - CHARACTER_JUMP_FORCE) / CHARACTER_MAX_VELOCITY_X
 
 # Wall scale
@@ -30,8 +31,11 @@ MIN_PLATFORM_SIZE: int = round(s.WIDTH * 0.234)
 MAX_PLATFORM_SIZE: int = round(s.WIDTH * 0.375)
 PLATFORM_THICKNESS: int = s.WIDTH * 0.02
 PLATFORM_DISTANCE: int = s.HEIGHT / 5
+
+# The starting y position of the first platform based on the display height
 PREV_Y: int = s.HEIGHT - (s.HEIGHT * 0.15)
-# Get platform_x
+
+# The formula below is to get the min and max values that the platform x can create based on the wall thickness so don't overlap
 MAX_HALF_SIZE: int = int(MAX_PLATFORM_SIZE / 2)
 PLATFORM_MIN_X: int = WALL_THICKNESS + MAX_HALF_SIZE
 PLATFORM_MAX_X: int = s.WIDTH - WALL_THICKNESS - MAX_HALF_SIZE
